+++ parsec/serde/__init__.py
@@ -6,7 +6,7 @@ from parsec.serde import fields
 from parsec.serde.exceptions import SerdeError, SerdeValidationError, SerdePackingError
 from parsec.serde.schema import BaseSchema, UnknownCheckedSchema, OneOfSchema, BaseCmdSchema
 from parsec.serde.packing import packb, unpackb
-from parsec.serde.porcelaine import BaseSerializer, MsgpackSerializer, ZipMsgpackSerializer
+from parsec.serde.serializer import BaseSerializer, MsgpackSerializer, ZipMsgpackSerializer
 
 __all__ = (
     "SerdeError",
